+++ cypress/e2e/it-system-usage/details-pages/it-system-usage.gdpr.cy.ts
@@ -1,4 +1,4 @@
-/// <reference types="Cypress" />
+import { TestRunner } from 'cypress/support/test-runner';
 
 const generalInformation = 'Generel information';
 const purposeInput = 'Systemets overordnede formål';
@@ -12,251 +12,180 @@ const userSupervisionAccordion = 'user-supervision-accordion';
 const riskAssessmentAccordion = 'risk-assessment-accordion';
 const datepickerToggle = 'datepicker-toggle';
 
+function setupTest() {
+  cy.requireIntercept();
+  cy.setupItSystemUsageIntercepts();
+  cy.intercept('/api/v2/it-system-usage-sensitive-personal-data-types?organizationUuid=*', {
+    fixture: './it-system-usage/it-system-usage-sensitive-personal-data-types.json',
+  });
+  cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/gdpr/it-system-usage-full-gdpr.json' });
+  cy.intercept('/api/v2/it-system-usage-registered-data-category-types?organizationUuid=*', {
+    fixture: './it-system-usage/it-system-usage-registered-data-category-types.json',
+  });
+  cy.setup(true, 'it-systems/it-system-usages');
+}
+
 describe('it-system-usage gdpr', () => {
-  beforeEach(() => {
-    cy.requireIntercept();
-    cy.setupItSystemUsageIntercepts();
-    cy.intercept('/api/v2/it-system-usage-sensitive-personal-data-types?organizationUuid=*', {
-      fixture: './it-system-usage/it-system-usage-sensitive-personal-data-types.json',
-    });
-    cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/gdpr/it-system-usage-full-gdpr.json' });
-    cy.intercept('/api/v2/it-system-usage-registered-data-category-types?organizationUuid=*', {
-      fixture: './it-system-usage/it-system-usage-registered-data-category-types.json',
-    });
-    cy.setup(true, 'it-systems/it-system-usages');
-    cy.contains('System 3').click();
-    cy.navigateToDetailsSubPage('GDPR');
-  });
+  const testRunner = new TestRunner(setupTest);
+  it('it-system-usage gdpr', () => {
+    testRunner.runTestWithSetup('can show GDPR tab and existing data in general input fields', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('patch');
 
-  it('can show GDPR tab and existing data in general input fields', () => {
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
+      cy.contains(generalInformation);
+      cy.contains('Yderligere information');
+      cy.input(purposeInput).should('have.value', 'Test purpose');
+      cy.dropdown(businessCriticalDropdown).should('have.text', 'Ja');
+      cy.dropdown(hostedAtDropdown).should('have.text', 'On-premise');
 
-    cy.contains(generalInformation);
-    cy.contains('Yderligere information');
-    cy.input(purposeInput).should('have.value', 'Test purpose');
-    cy.dropdown(businessCriticalDropdown).should('have.text', 'Ja');
-    cy.dropdown(hostedAtDropdown).should('have.text', 'On-premise');
-
-    verifyLinkTextbox('directory-documentation-link', 'newName: newUrl');
-  });
-
-  it('can edit purpose', () => {
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
-    const newPurpose = 'New purpose';
-    cy.input(purposeInput).clear().type(newPurpose);
-    cy.contains(generalInformation).click();
-
-    cy.verifyRequestUsingDeepEq('patch', 'request.body', { gdpr: { purpose: newPurpose } });
-    cy.input(purposeInput).should('have.value', newPurpose);
-  });
-
-  it('can edit business critical status', () => {
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
-    const newBusinessCritical = 'Nej';
-    cy.dropdown(businessCriticalDropdown, newBusinessCritical, true);
-    cy.contains(generalInformation).click();
-
-    verifyGdprPatchRequest({ businessCritical: 'No' });
-    cy.dropdown(businessCriticalDropdown).should('have.text', newBusinessCritical);
-  });
-
-  it('can edit hosted at status', () => {
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
-    const newHostedAt = 'Eksternt';
-    cy.dropdown(hostedAtDropdown, newHostedAt, true);
-    cy.contains(generalInformation).click();
-
-    verifyGdprPatchRequest({ hostedAt: 'External' });
-    cy.dropdown(hostedAtDropdown).should('have.text', newHostedAt);
-  });
-
-  it('can edit data sensitivity levels', () => {
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
-    cy.getByDataCy(dataSensitivityAccordion).click();
-
-    cy.input('Straffedomme og lovovertrædelser').should('be.checked');
-    cy.input(personDataCheckbox).should('not.be.checked');
-    cy.input(personDataCheckbox).click();
-    verifyGdprPatchRequest({ dataSensitivityLevels: ['PersonData', 'LegalData'] });
-  });
-
-  it('can edit specific personal data, and sub-options when main option is selected', () => {
-    cy.getByDataCy(dataSensitivityAccordion).click();
-    const nestedCheckboxes = ['CPR-nr.', 'Væsentlige sociale problemer', 'Andre rent private forhold'];
-    nestedCheckboxes.forEach((checkBox) => {
-      cy.input(checkBox).should('be.disabled');
-    });
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-full-gdpr.json',
-    }).as('patch');
-
-    cy.input('Almindelige personoplysninger').click();
-    verifyGdprPatchRequest({ dataSensitivityLevels: ['PersonData', 'LegalData'] });
-
-    nestedCheckboxes.forEach((checkBox) => {
-      cy.input(checkBox).should('be.enabled');
+      verifyLinkTextbox('directory-documentation-link', 'newName: newUrl');
     });
 
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patchSpecificPersonalData');
-    cy.input('CPR-nr.').click({ force: true });
-    cy.verifyRequestUsingDeepEq('patchSpecificPersonalData', 'request.body', {
-      gdpr: { specificPersonalData: ['CprNumber'] },
-    });
-  });
+    testRunner.runTestWithSetup('can edit purpose', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('purposePatch');
 
-  it('can edit sensitive personal data, and sub-options when main option is selected', () => {
-    cy.getByDataCy(dataSensitivityAccordion).click();
-    const nestedCheckboxes = ['data type 1', 'data type 2'];
-    nestedCheckboxes.forEach((checkBox) => {
-      cy.input(checkBox).should('be.disabled');
-    });
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-gdpr-sensitive-data.json',
-    }).as('patch');
+      const newPurpose = 'New purpose';
+      cy.input(purposeInput).clear().type(newPurpose);
+      cy.contains(generalInformation).click();
 
-    cy.input('Følsomme personoplysninger').click();
-    verifyGdprPatchRequest({ dataSensitivityLevels: ['SensitiveData', 'LegalData'] });
-
-    nestedCheckboxes.forEach((checkBox) => {
-      cy.input(checkBox).should('be.enabled');
+      cy.verifyRequestUsingDeepEq('purposePatch', 'request.body', { gdpr: { purpose: newPurpose } });
+      cy.input(purposeInput).should('have.value', newPurpose);
     });
 
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patchSensitivePersonalData');
-    cy.input('data type 1').click({ force: true });
-    verifyGdprPatchRequest(
-      { sensitivePersonDataUuids: ['00000000-0000-0000-0000-000000000000'] },
-      'patchSensitivePersonalData',
-    );
-  });
+    testRunner.runTestWithSetup('can edit business critical status', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('businessCriticalPatch');
+      const newBusinessCritical = 'Nej';
+      cy.dropdown(businessCriticalDropdown, newBusinessCritical, true);
+      cy.contains(generalInformation).click();
 
-  it('can edit registered categories of data', () => {
-    cy.getByDataCy(registedCategoriesAccordion).click();
-    const checkBox = 'data category 1';
+      verifyGdprPatchRequest({ businessCritical: 'No' }, 'businessCriticalPatch');
+      cy.dropdown(businessCriticalDropdown).should('have.text', newBusinessCritical);
+    });
 
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('patch');
-    cy.input(checkBox).click();
-    verifyGdprPatchRequest({ registeredDataCategoryUuids: ['00000000-0000-0000-0000-000000000000'] });
-  });
+    testRunner.runTestWithSetup('can edit hosted at status', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('hostedAtPatch');
+      const newHostedAt = 'Eksternt';
+      cy.dropdown(hostedAtDropdown, newHostedAt, true);
+      cy.contains(generalInformation).click();
 
-  it('can edit technical precautions', () => {
-    cy.getByDataCy(technicalPrecautionsAccordion)
-      .click()
-      .within(() => {
-        cy.getByDataCy('technical-precautions-dropdown').should('contain', 'Nej');
-        const nestedCheckboxes = ['Kryptering', 'Pseudonomisering', 'Adgangsstyring', 'Logning'];
+      verifyGdprPatchRequest({ hostedAt: 'External' }, 'hostedAtPatch');
+      cy.dropdown(hostedAtDropdown).should('have.text', newHostedAt);
+    });
+
+    testRunner.runTestWithSetup('can edit data sensitivity levels', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('sensitivityLevelsPatch');
+      cy.getByDataCy(dataSensitivityAccordion).click();
+
+      cy.input('Straffedomme og lovovertrædelser').should('be.checked');
+      cy.input(personDataCheckbox).should('not.be.checked');
+      cy.input(personDataCheckbox).click();
+      verifyGdprPatchRequest({ dataSensitivityLevels: ['PersonData', 'LegalData'] }, 'sensitivityLevelsPatch');
+    });
+
+    testRunner.runTestWithSetup('can edit specific personal data, and sub-options when main option is selected', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.getByDataCy(dataSensitivityAccordion).click();
+      const nestedCheckboxes = ['CPR-nr.', 'Væsentlige sociale problemer', 'Andre rent private forhold'];
+      nestedCheckboxes.forEach((checkBox) => {
+        cy.input(checkBox).should('be.disabled');
+      });
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-full-gdpr.json',
+      }).as('personalDataPatch');
+
+      cy.input('Almindelige personoplysninger').click();
+      verifyGdprPatchRequest({ dataSensitivityLevels: ['PersonData', 'LegalData'] }, 'personalDataPatch');
+
+      nestedCheckboxes.forEach((checkBox) => {
+        cy.input(checkBox).should('be.enabled');
+      });
+
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('patchSpecificPersonalData');
+      cy.input('CPR-nr.').click({ force: true });
+      cy.verifyRequestUsingDeepEq('patchSpecificPersonalData', 'request.body', {
+        gdpr: { specificPersonalData: ['CprNumber'] },
+      });
+    });
+
+    testRunner.runTestWithSetup(
+      'can edit sensitive personal data, and sub-options when main option is selected',
+      () => {
+        cy.contains('System 3').click();
+        cy.navigateToDetailsSubPage('GDPR');
+        cy.getByDataCy(dataSensitivityAccordion).click();
+        const nestedCheckboxes = ['data type 1', 'data type 2'];
         nestedCheckboxes.forEach((checkBox) => {
           cy.input(checkBox).should('be.disabled');
         });
-
         cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-          fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-        }).as('patchYesToApplied');
-        cy.getByDataCy('technical-precautions-dropdown').click().contains('Ja').click({ force: true });
-        verifyGdprPatchRequest({ technicalPrecautionsInPlace: 'Yes' }, 'patchYesToApplied');
+          fixture: './it-system-usage/gdpr/it-system-usage-gdpr-sensitive-data.json',
+        }).as('pesonalDataPatch2');
+
+        cy.input('Følsomme personoplysninger').click();
+        verifyGdprPatchRequest({ dataSensitivityLevels: ['SensitiveData', 'LegalData'] }, 'pesonalDataPatch2');
+
         nestedCheckboxes.forEach((checkBox) => {
           cy.input(checkBox).should('be.enabled');
         });
 
         cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
           fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-        }).as('patchAddPrecaution');
-        cy.input('Kryptering').click();
-        verifyGdprPatchRequest({ technicalPrecautionsApplied: ['Encryption'] }, 'patchAddPrecaution');
-        verifyLinkTextAndPressEdit('technical-precautions-documentation-link', 'newName: newUrl');
-      });
-    verifyLinkEditDialogAndPerformEdit();
-  });
+        }).as('patchSensitivePersonalData');
+        cy.input('data type 1').click({ force: true });
+        verifyGdprPatchRequest(
+          { sensitivePersonDataUuids: ['00000000-0000-0000-0000-000000000000'] },
+          'patchSensitivePersonalData'
+        );
+      }
+    );
 
-  it('can edit user supervision', () => {
-    cy.getByDataCy(userSupervisionAccordion)
-      .click()
-      .within(() => {
-        const urlSectionDropdown = 'url-section-dropdown';
-        cy.getByDataCy(urlSectionDropdown).should('contain', 'Nej');
-        cy.get(datepickerToggle).should('not.exist');
+    testRunner.runTestWithSetup('can edit registered categories of data', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      cy.getByDataCy(registedCategoriesAccordion).click();
+      const checkBox = 'data category 1';
 
-        cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-          fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-        }).as('patchYesToSupervision');
-        cy.getByDataCy(urlSectionDropdown).click().contains('Ja').click({ force: true });
-        verifyGdprPatchRequest({ userSupervision: 'Yes' }, 'patchYesToSupervision');
-
-        cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-          fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-        }).as('patchAddPrecaution');
-        cy.getByDataCy('base-date-url-section-selector').type('2024-02-15');
-
-        verifyLinkTextAndPressEdit('base-date-url-section-selector', 'newName: newUrl');
-      });
-    verifyLinkEditDialogAndPerformEdit();
-  });
-
-  it('can edit risk assessment', () => {
-    cy.getByDataCy(riskAssessmentAccordion).click();
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('riskAssessmentPatch');
-    const date = '15-02-2024';
-
-    cy.dropdownByCy('risk-assessment-dropdown', 'Ja', true);
-    verifyGdprPatchRequest({ riskAssessmentConducted: 'Yes' }, 'riskAssessmentPatch');
-
-    cy.inputByCy('planned-date-datepicker').type(date);
-    verifyAppNotification();
-    cy.getByDataCy('conducted-date-datepicker').type(date);
-    verifyAppNotification();
-    cy.dropdownByCy('assessment-result-dropdown', 'Lav risiko', true);
-    verifyAppNotification();
-    cy.getByDataCy('assessment-notes').type('Test');
-    verifyAppNotification();
-
-    verifyLinkTextAndPressEdit('risk-documentation-link', 'newName: newUrl');
-    verifyLinkEditDialogAndPerformEdit();
-  });
-
-  it('cannot save directory documentation with unchanged url', () => {
-    verifyLinkTextAndPressEdit('directory-documentation-link', 'newName: newUrl');
-
-    cy.get('app-edit-url-dialog').within(() => {
-      cy.verifyDialogConfirmButtonDisabledByReactiveForm('edit-url-save-button');
-    });
-  });
-
-  it('can edit retention period', () => {
-    cy.getByDataCy('retention-period-accordion').click();
-
-    cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
-    }).as('retentionPeriodPatch');
-
-    cy.dropdownByCy('retention-period-dropdown', 'Ja', true);
-    verifyAppNotification();
-
-    cy.clock().then((clock) => {
-      clock.setSystemTime(new Date('10/10/2022'));
+      cy.intercept('PATCH', '/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/gdpr/it-system-usage-updated-gdpr.json',
+      }).as('categoriesPatch');
+      cy.input(checkBox).click();
+      verifyGdprPatchRequest(
+        { registeredDataCategoryUuids: ['00000000-0000-0000-0000-000000000000'] },
+        'categoriesPatch'
+      );
     });
 
-    cy.datepickerByCy('retention-period-datepicker', '15');
-    verifyAppNotification();
+    testRunner.runTestWithSetup('cannot save directory documentation with unchanged url', () => {
+      cy.contains('System 3').click();
+      cy.navigateToDetailsSubPage('GDPR');
+      verifyLinkTextAndPressEdit('directory-documentation-link', 'newName: newUrl');
 
-    cy.inputByCy('retention-period-numeric-input').type('20', { force: true });
-    verifyAppNotification();
+      cy.get('app-edit-url-dialog').within(() => {
+        cy.verifyDialogConfirmButtonDisabledByReactiveForm('edit-url-save-button');
+      });
+    });
   });
 });
