+++ cypress/e2e/it-system-usage/details-pages/it-system-usage.kle.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="Cypress" />
-
 import { TestRunner } from 'cypress/support/test-runner';
 
 function setupTest() {
