+++ cypress/e2e/it-system-usage/details-pages/it-system-usage.system-relations.cy.ts
@@ -1,4 +1,4 @@
-/// <reference types="Cypress" />
+import { TestRunner } from 'cypress/support/test-runner';
 
 interface RelationRow {
   systemUsageName: string;
@@ -11,97 +11,102 @@ interface RelationRow {
   expectedInvalidUrlReference?: string;
 }
 
+function setupTest() {
+  cy.requireIntercept();
+  cy.setupItSystemUsageIntercepts();
+
+  cy.setup(true, 'it-systems/it-system-usages', './shared/it-system-usage-ui-customization-no-gdpr-and-lifecycle.json');
+}
+
 describe('it-system-usage system-relations', () => {
-  beforeEach(() => {
-    cy.requireIntercept();
-    cy.setupItSystemUsageIntercepts();
-    cy.setup(true, 'it-systems/it-system-usages');
+  const testRunner = new TestRunner(setupTest);
+
+  it('it-system-usage system-relations', () => {
+    testRunner.runTestWithSetup('can show empty Relations', () => {
+      cy.intercept('/api/v2/it-system-usages/*', {
+        fixture: './it-system-usage/relations/it-system-usage-no-relations.json',
+      });
+
+      cy.contains('System 3').click();
+
+      cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', []);
+      cy.intercept('/api/v2/*it-system-usage-relation-frequency-types*', []);
+
+      cy.navigateToDetailsSubPage('Relationer');
+
+      cy.getCardWithTitle("'System 3' har følgende relationer til andre systemer/snitflader").within(() => {
+        cy.contains("Endnu ingen relationer tilføjet fra 'System 3' til andre systemer/snitflader");
+        cy.contains('Opret relation');
+      });
+
+      cy.getCardWithTitle("Andre systemer har følgende relationer til 'System 3'").within(() => {
+        cy.contains("Ingen relationer fra andre systemer til 'System 3' fundet");
+      });
+    });
+
+    testRunner.runTestWithSetup('can show Relations', () => {
+      cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
+
+      cy.contains('System 3').click();
+
+      cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', {
+        fixture: './it-system-usage/relations/it-system-usage-incoming-relations.json',
+      });
+      cy.intercept('/api/v2/*it-system-usage-relation-frequency-types*', {
+        fixture: './it-system-usage/relations/relation-frequency-types.json',
+      });
+
+      cy.navigateToDetailsSubPage('Relationer');
+
+      const expectedOutgoingRows = [
+        {
+          systemUsageName: 'Aplanner',
+          relationInterfaceName: 'Lorem ipsum',
+          associatedContractName: 'Dolor sit',
+          relationFrequencyName: 'Kvartal',
+          isFrequencyObsolete: false,
+          description: 'Sed at nibh ac tellus tempor dignissim quis id urna',
+          expectedValidUrlReference: 'https://www.google.com',
+        },
+        {
+          systemUsageName: 'Vestibulum',
+          relationInterfaceName: 'Quisque nec',
+          associatedContractName: 'erat volutpat',
+          relationFrequencyName: 'Kvartal2',
+          isFrequencyObsolete: true,
+          description:
+            'Nulla non quam rhoncus lacus ultricies interdum ac ut dui. Duis quis egestas arcu. Integer ultrices turpis nec felis lobortis, in pretium erat lobortis',
+          expectedInvalidUrlReference: 'www.google.com',
+        },
+      ];
+
+      const expectedIncomingRows = [
+        {
+          systemUsageName: 'testUsage',
+          relationInterfaceName: 'testInterface',
+          associatedContractName: 'DefaultTestItContract',
+          relationFrequencyName: 'Kvartal',
+          isFrequencyObsolete: false,
+          description: 'test description',
+          expectedValidUrlReference: 'https://www.google.com',
+        },
+        {
+          systemUsageName: 'testUsage2',
+          relationInterfaceName: 'testInterface2',
+          associatedContractName: 'DefaultTestItContract2',
+          relationFrequencyName: 'Kvartal2',
+          isFrequencyObsolete: true,
+          description: 'test description2',
+          expectedInvalidUrlReference: 'www.google.com',
+        },
+      ];
+
+      verifyRelationTable('System 3', expectedOutgoingRows, true);
+      verifyRelationTable('System 3', expectedIncomingRows, false);
+    });
   });
-
-  it('can show empty Relations', () => {
-    cy.intercept('/api/v2/it-system-usages/*', {
-      fixture: './it-system-usage/relations/it-system-usage-no-relations.json',
-    });
-
-    cy.contains('System 3').click();
-
-    cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', []);
-    cy.intercept('/api/v2/*it-system-usage-relation-frequency-types*', []);
-
-    cy.navigateToDetailsSubPage('Relationer');
-
-    cy.getCardWithTitle("'System 3' har følgende relationer til andre systemer/snitflader").within(() => {
-      cy.contains("Endnu ingen relationer tilføjet fra 'System 3' til andre systemer/snitflader");
-      cy.contains('Opret relation');
-    });
-
-    cy.getCardWithTitle("Andre systemer har følgende relationer til 'System 3'").within(() => {
-      cy.contains("Ingen relationer fra andre systemer til 'System 3' fundet");
-    });
-  });
-
-  it('can show Relations', () => {
-    cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
-
-    cy.contains('System 3').click();
-
-    cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', {
-      fixture: './it-system-usage/relations/it-system-usage-incoming-relations.json',
-    });
-    cy.intercept('/api/v2/*it-system-usage-relation-frequency-types*', {
-      fixture: './it-system-usage/relations/relation-frequency-types.json',
-    });
-
-    cy.navigateToDetailsSubPage('Relationer');
-
-    const expectedOutgoingRows = [
-      {
-        systemUsageName: 'Aplanner',
-        relationInterfaceName: 'Lorem ipsum',
-        associatedContractName: 'Dolor sit',
-        relationFrequencyName: 'Kvartal',
-        isFrequencyObsolete: false,
-        description: 'Sed at nibh ac tellus tempor dignissim quis id urna',
-        expectedValidUrlReference: 'https://www.google.com',
-      },
-      {
-        systemUsageName: 'Vestibulum',
-        relationInterfaceName: 'Quisque nec',
-        associatedContractName: 'erat volutpat',
-        relationFrequencyName: 'Kvartal2',
-        isFrequencyObsolete: true,
-        description:
-          'Nulla non quam rhoncus lacus ultricies interdum ac ut dui. Duis quis egestas arcu. Integer ultrices turpis nec felis lobortis, in pretium erat lobortis',
-        expectedInvalidUrlReference: 'www.google.com',
-      },
-    ];
-
-    const expectedIncomingRows = [
-      {
-        systemUsageName: 'testUsage',
-        relationInterfaceName: 'testInterface',
-        associatedContractName: 'DefaultTestItContract',
-        relationFrequencyName: 'Kvartal',
-        isFrequencyObsolete: false,
-        description: 'test description',
-        expectedValidUrlReference: 'https://www.google.com',
-      },
-      {
-        systemUsageName: 'testUsage2',
-        relationInterfaceName: 'testInterface2',
-        associatedContractName: 'DefaultTestItContract2',
-        relationFrequencyName: 'Kvartal2',
-        isFrequencyObsolete: true,
-        description: 'test description2',
-        expectedInvalidUrlReference: 'www.google.com',
-      },
-    ];
-
-    verifyRelationTable('System 3', expectedOutgoingRows, true);
-    verifyRelationTable('System 3', expectedIncomingRows, false);
-  });
-
   it('can add Relation', () => {
+    setupTest();
     cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
 
     cy.contains('System 3').click();
@@ -135,7 +140,31 @@ describe('it-system-usage system-relations', () => {
     cy.contains('Relation tilføjet');
   });
 
+  it('can delete Relation', () => {
+    setupTest();
+    cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
+
+    cy.contains('System 3').click();
+
+    cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', []);
+    cy.intercept('/api/v2/it-system-usage-relation-frequency-types*', []);
+
+    cy.navigateToDetailsSubPage('Relationer');
+
+    cy.getRowForElementContent('Aplanner').get('app-trashcan-icon').first().click({ force: true });
+
+    cy.verifyYesNoConfirmationDialogAndConfirm(
+      'DELETE',
+      '**/system-relations/*',
+      {},
+      'Er du sikker på at du vil fjerne denne relation'
+    );
+
+    cy.contains('Relationen er slettet');
+  });
+
   it('can modify Relation', () => {
+    setupTest();
     cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
 
     cy.contains('System 3').click();
@@ -169,33 +198,11 @@ describe('it-system-usage system-relations', () => {
     cy.contains('Relation ændret');
   });
 
-  it('can delete Relation', () => {
-    cy.intercept('/api/v2/it-system-usages/*', { fixture: './it-system-usage/it-system-usage.json' });
-
-    cy.contains('System 3').click();
-
-    cy.intercept('/api/v2/it-system-usages/*/incoming-system-relations', []);
-    cy.intercept('/api/v2/it-system-usage-relation-frequency-types*', []);
-
-    cy.navigateToDetailsSubPage('Relationer');
-
-    cy.getRowForElementContent('Aplanner').get('app-trashcan-icon').first().click({ force: true });
-
-    cy.verifyYesNoConfirmationDialogAndConfirm(
-      'DELETE',
-      '**/system-relations/*',
-      {},
-      'Er du sikker på at du vil fjerne denne relation',
-    );
-
-    cy.contains('Relationen er slettet');
-  });
-
   function verifyRelationTable(systemName: string, rows: RelationRow[], isOutgoing: boolean) {
     cy.getCardWithTitle(
       isOutgoing
         ? `'${systemName}' har følgende relationer til andre systemer/snitflader`
-        : `Andre systemer har følgende relationer til '${systemName}'`,
+        : `Andre systemer har følgende relationer til '${systemName}'`
     ).within(() => {
       for (const expectedRow of rows) {
         const row = () => cy.getRowForElementContent(expectedRow.systemUsageName);
